+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Types of changes: Added, Changed, Deprecated, Removed, Fixed, Security
 
+## [0.2.0] - 2026-03-24
+
+### Added
+
+- LOCALNET/WHITELISTED classification
+
+### Fixed
+
+- Docker bypass duplicates
+- addrtype catch-all false positives
+
 ## [0.1.0] - 2026-03-24
 
 ### Added